+++ src/sqlitely/conf.py
@@ -30,7 +30,7 @@ except ImportError: wx = None
 
 """Program title, version number and version date."""
 Title = "SQLitely"
-Version = "2.2.dev56"
+Version = "2.2.dev57"
 VersionDate = "08.10.2023"
 
 if getattr(sys, "frozen", False):